requirements.txt
streamlit>=1.18
google-generativeai
